[SharpFPS] Various Improvements and Fixes
[SharpFPS] Consider lack of sleep!
[SharpFPS] Implement "Single Point Frame Pacing"

Single point frame pacing is a modification of the original SharpFPS that allows you to limit the framerate by basically placing a call to SharpFPS at a single place within the recurring logic loop, as opposed to wrapping the start and end of recurring loop with StartFrame and EndFrame.

This is to allow for frame pacing in cases where there is external code outside of the user's control, where a StartFrame or EndFrame cannot be used.

[SharpFPS]
SharpFPS now is more accurate and compensates for the time overslept, which was previously causing frames to last slightly less less than the cap was actually set to.

A bug in SharpFPS causing the framerate to be incorrectly reported under a specific circumstance has been fixed.

Calculation of average FPS has been improved, using a default of 100 points of polling to calculate the actual average.

That said, while the framerate is now reported correctly, there is still an issue. The actual framerate is lesser than the framerate cap! (The reported framerate by SharpFPS, and external utilities is correct though)

I'm pretty sure this is because RenderLoop.Run() does not include the time taken to pain the actual panel, thus that is not compensated for. I'll see what can be done about it.